+++ CDiscretizedDistribution.h
@@ -1,7 +1,8 @@
 /***************************************************************************************************
 *  The PolyMoSim project is distributed under the following license:
 *  
-*  Copyright (c) 2006-2022, Christoph Mayer, Forschungsmuseum Alexander Koenig, Bonn, Germany
+*  Copyright (c) 2006-2025, Christoph Mayer, Leibniz Institute for the Analysis of Biodiversity Change,
+*  Bonn, Germany
 *  All rights reserved.
 *  
 *  Redistribution and use in source and binary forms, with or without
@@ -55,7 +56,7 @@
 #include <iostream>
 #include "math_expression_parser.h"
 #include <vector>
-#include "faststring2.h"
+#include "faststring3.h"
 #include <algorithm>
 #include "CRandom.h"
 
@@ -146,7 +147,7 @@ class DiscretizedDistribution
     }
   }
 
-  double get_coordinate(unsigned i)
+  double get_coordinate(unsigned long i)
   {
     return a+step*(i+0.5);
   }
